Rails integration
Added a Railtie to enable configuration-less integration with Rails.

Changed the configuration object to be a singleton that is shared across all
calls to Hatchet.configure rather than replacing the current configuration.

To that end a reset! method was added to the configuration object that resets
the levels and removes all of the appenders.

The version was incremented to 0.0.7 in preparation for the next release.