Improved Rails integration
Now picks up the logging level of the Rails logger and uses that as the default
level for Hatchet.

The configuration object can now take a block to its configure method to make
larger blocks of configuration nicer to read in Rails environment configs.

Added the ability to define what the default formatter is rather than it having
to be a StandardFormatter.