Add the ability to pass an error through
This commit adds the ability to provide an error along with a message to each
level of the logger. This is made available to appenders and their formatters
through the now publicly exposed Message class via the Message#error attribute.

No changes have been made to the Plain-, Simple-, or StandardFormatter at this
point to utilize the Message#error attribute.